rebuild-inputs/evergreen-tvevents/adjacent/rebuilder-evergreen-template-repo-python/pyproject.toml
[project]
name = "demo-app"
version = "1.0.0"
description = "Inscape Demo application."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "async-timeout==5.0.1",
    "attrs==24.2.0",
    "boto==2.49.0",
    "boto3==1.38.14",
    "botocore==1.38.14",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "fakeredis==2.29.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-api-core==1.34.1",
    "google-auth==2.40.1",
    "googleapis-common-protos==1.70.0",

    "idna==3.10",
    "importlib-metadata>=3.6",

    "jmespath==1.0.1",
    "jsonschema==3.2.0",

    "packaging==24.2",
    "protobuf==3.20.3",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pygerduty==0.38.3",
    "pymemcache==4.0.0",
    "PyMySQL==1.1.1",
    "pyrsistent==0.20.0",
    "python-consul==1.1.0",
    "python-dateutil==2.9.0.post0",
    "PyYAML==6.0.2",
    "pyzmq==26.4.0",
    "redis==6.0.0",
    "requests>=2.32.4",
    "rsa==4.9.1",
    "s3transfer==0.12.0",
    "schema==0.7.7",
    "six==1.17.0",
    "sortedcontainers==2.4.0",
    "typing_extensions==4.13.2",
    "urllib3==1.26.20",

    "opentelemetry-api==1.31.1",
    "opentelemetry-distro==0.52b1",
    "opentelemetry-exporter-otlp-proto-grpc==1.15.0",
    "opentelemetry-instrumentation==0.52b1",
    "opentelemetry-instrumentation-asyncio==0.52b1",
    "opentelemetry-instrumentation-boto==0.52b1",
    "opentelemetry-instrumentation-boto3sqs==0.52b1",
    "opentelemetry-instrumentation-botocore==0.52b1",
    "opentelemetry-instrumentation-click==0.52b1",
    "opentelemetry-instrumentation-dbapi==0.52b1",
    "opentelemetry-instrumentation-fastapi==0.52b1",
    "opentelemetry-instrumentation-grpc==0.52b1",
    "opentelemetry-instrumentation-jinja2==0.52b1",
    "opentelemetry-instrumentation-logging==0.52b1",
    "opentelemetry-instrumentation-psycopg2==0.52b1",
    "opentelemetry-instrumentation-pymemcache==0.52b1",
    "opentelemetry-instrumentation-pymysql==0.52b1",
    "opentelemetry-instrumentation-redis==0.52b1",
    "opentelemetry-instrumentation-requests==0.52b1",
    "opentelemetry-instrumentation-sqlite3==0.52b1",
    "opentelemetry-instrumentation-threading==0.52b1",
    "opentelemetry-instrumentation-urllib==0.52b1",
    "opentelemetry-instrumentation-urllib3==0.52b1",
    "opentelemetry-instrumentation-asgi==0.52b1",
    "opentelemetry-propagator-aws-xray==1.0.2",
    "opentelemetry-proto==1.15.0",
    "opentelemetry-sdk==1.31.1",
    "opentelemetry-semantic-conventions==0.52b1",
    "opentelemetry-util-http==0.52b1",
    "opentelemetry-exporter-otlp-proto-http",
]

[project.optional-dependencies]
dev = [
    "validate-pyproject==0.18",
    "pytest-pythonpath==0.7.4",
    "pytest-cov==6.2.1",
    "complexipy==3.3.0",
    "pip-tools==7.4.1",
    "pytest==6.2.5",
    "pylint==3.2.5",
    "black==24.4.2",
    "toml==0.10.2",
    "pytest-env",
    "coverage[toml]==7.10.1",
    "typer[all]==0.15.4",
    "mypy",
    "types-jsonschema",
    "types-psycopg2",
    "types-requests",
    "pip==24.3.1",
    "setuptools==78.1.1"
]

[build-system]
requires = ["setuptools>=78"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"app" = "app"}
py-modules = ["app"]

[tool.setuptools.packages.find]
where = ["app"]
namespaces = false

[tool.black]
line-length = 88
target-version = ['py312']
skip-string-normalization = true

[tool.mypy]
ignore_missing_imports = true